+++ frontend/src/components/offer/OfferDetailsItem.js
@@ -1,4 +1,5 @@
-import { getDeliveryDay, convertToEuro } from "../../services/helpers";
+import convertToEuro from "../../services/convertToEuro";
+import getDeliveryDay from "../../services/getDeliveryDay";
 import PropTypes from "prop-types";
 import React from "react";
 import styled from "styled-components";
