+++ frontend/src/services/checkOnStock.js
@@ -1,0 +1,5 @@
+export default function checkOnStock(offersData) {
+  return offersData.some((offer) => offer.onStock === true)
+    ? "Auf Lager."
+    : "Leider nicht verfügbar.";
+}
