+++ frontend/src/services/convertToEuro.js
@@ -1,0 +1,7 @@
+export default function convertToEuro(number) {
+  const convertNumber = new Intl.NumberFormat("de-DE", {
+    style: "currency",
+    currency: "EUR",
+  }).format(number);
+  return convertNumber;
+}
