+++ frontend/src/services/findProduct.js
@@ -1,0 +1,5 @@
+export default function findProduct(id, products) {
+  let foundProduct = products.find((elem) => elem.id == id);
+  console.log(foundProduct);
+  return foundProduct;
+}
