+++ frontend/src/services/getCartImage.js
@@ -1,0 +1,4 @@
+export default function getCartImage(productData, cartId) {
+  const foundIndex = productData.findIndex((elem) => elem.id === cartId);
+  return productData[foundIndex].image;
+}
