+++ frontend/src/services/getFastestDeliveryDay.js
@@ -1,0 +1,16 @@
+import getDeliveryDay from "./getDeliveryDay";
+
+export default function getFastestDeliveryDay(offers) {
+  function getFastestDelivery(offersData) {
+    const offers = [];
+    offersData.forEach((offer) => {
+      offers.push(offer.deliveryTime);
+    });
+    return Math.min.apply(Math, offers);
+  }
+
+  const deliveryTime = getFastestDelivery(offers);
+  const fastestDeliveryday = getDeliveryDay(deliveryTime);
+
+  return fastestDeliveryday;
+}
